# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WedWithMe REST API",
    "version": "1.0.0",
    "description": "Autogenerated OpenAPI specification for the WedWithMe Next.js Backend."
  },
  "servers": [
    {
      "url": "/api",
      "description": "Local API Environment"
    }
  ],
  "paths": {
...
    "/admin/analytics/financials": {
      "get": {
        "summary": "GET /admin/analytics/financials",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/analytics/performance": {
      "get": {
        "summary": "GET /admin/analytics/performance",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/cancellations/rules": {
      "get": {
        "summary": "GET /admin/cancellations/rules",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "summary": "POST /admin/cancellations/rules",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/admin/disputes/{id}": {
      "get": {
        "summary": "GET /admin/disputes/{id}",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "patch": {
        "summary": "PATCH /admin/disputes/{id}",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/fraud/duplicates": {
      "get": {
        "summary": "GET /admin/fraud/duplicates",
        "tags": [
          "Admin"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}